refactor: create app form to package
