+++ apps/nextjs/src/app/[locale]/manage/apps/new/page.tsx
@@ -3,9 +3,9 @@ import { Container, Stack, Title } from "@mantine/core";
 
 import { auth } from "@homarr/auth/next";
 import { getI18n } from "@homarr/translation/server";
+import { AppNewForm } from "@homarr/forms-collection";
 
 import { DynamicBreadcrumb } from "~/components/navigation/dynamic-breadcrumb";
-import { AppNewForm } from "../../../../../../../../packages/ui/src/components/forms/new-app/_app-new-form";
 
 export default async function AppNewPage() {
   const session = await auth();
