+++ apps/nextjs/tsconfig.json
@@ -12,6 +12,6 @@
     ],
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
   },
-  "include": [".", ".next/types/**/*.ts", "../../packages/ui/src/components/forms/new-app/_form.tsx", "../../packages/ui/src/components/forms/new-app/_app-new-form.tsx"],
+  "include": [".", ".next/types/**/*.ts"],
   "exclude": ["node_modules", ".next"]
 }
