+++ packages/forms-collection/src/icon-picker/icon-picker.module.css
