+++ packages/forms-collection/src/new-app/_form.tsx
@@ -8,12 +8,12 @@ import type { z } from "zod";
 import { useZodForm } from "@homarr/form";
 import { useI18n } from "@homarr/translation/client";
 import { validation } from "@homarr/validation";
-
-import { IconPicker } from "~/components/icons/picker/icon-picker";
+import { IconPicker } from "../icon-picker/icon-picker";
 
 type FormType = z.infer<typeof validation.app.manage>;
 
 interface AppFormProps {
+  showBackToOverview: boolean;
   buttonLabels: {
     submit: string;
     submitAndCreateAnother?: string;
@@ -24,11 +24,12 @@ interface AppFormProps {
 }
 
 export const AppForm = ({
-  buttonLabels,
-  handleSubmit: originalHandleSubmit,
-  initialValues,
-  isPending,
-}: AppFormProps) => {
+                          buttonLabels,
+                          showBackToOverview,
+                          handleSubmit: originalHandleSubmit,
+                          initialValues,
+                          isPending,
+                        }: AppFormProps) => {
   const t = useI18n();
 
   const form = useZodForm(validation.app.manage, {
@@ -45,9 +46,9 @@ export const AppForm = ({
     const redirect = !shouldCreateAnother.current;
     const afterSuccess = shouldCreateAnother.current
       ? () => {
-          form.reset();
-          shouldCreateAnother.current = false;
-        }
+        form.reset();
+        shouldCreateAnother.current = false;
+      }
       : undefined;
     originalHandleSubmit(values, redirect, afterSuccess);
   };
@@ -55,15 +56,16 @@ export const AppForm = ({
   return (
     <form onSubmit={form.onSubmit(handleSubmit)}>
       <Stack>
-        <TextInput {...form.getInputProps("name")} withAsterisk label={t("app.field.name.label")} />
+        <TextInput {...form.getInputProps("name")} withAsterisk label={t("app.field.name.label")}/>
         <IconPicker {...form.getInputProps("iconUrl")} />
-        <Textarea {...form.getInputProps("description")} label={t("app.field.description.label")} />
-        <TextInput {...form.getInputProps("href")} label={t("app.field.url.label")} />
+        <Textarea {...form.getInputProps("description")} label={t("app.field.description.label")}/>
+        <TextInput {...form.getInputProps("href")} label={t("app.field.url.label")}/>
 
         <Group justify="end">
-          <Button variant="default" component={Link} href="/manage/apps">
+
+          {showBackToOverview && (<Button variant="default" component={Link} href="/manage/apps">
             {t("common.action.backToOverview")}
-          </Button>
+          </Button>)}
           {buttonLabels.submitAndCreateAnother && (
             <Button
               type="submit"
