+++ packages/modals-collection/src/apps/quick-add-app/quick-add-app-modal.tsx
@@ -1,9 +1,10 @@
 import { createModal } from "@homarr/modals";
+import { AppNewForm } from "@homarr/forms-collection";
 
 export const QuickAddAppModal = createModal(({ actions }) => {
-    return <AppNewForm />;
+  return <AppNewForm showCreateAnother={false} showBackToOverview={false}/>;
 }).withOptions({
-    defaultTitle(t) {
-        return 'TEST'; // TODO: change
-    },
+  defaultTitle(t) {
+    return 'TEST'; // TODO: change
+  },
 });
